net: add contentProtections informations to parsed manifests
