+++ src/parsers/manifest/dash/node_parsers/ContentProtection.ts
@@ -17,8 +17,7 @@
 export interface IParsedContentProtection {
   schemeIdUri?: string;
   value?: string;
-  KID?: string;
-  pssh?: string;
+  keyId?: string;
 }
 
 /**
@@ -30,36 +29,35 @@ export interface IParsedContentProtection {
 export default function parseContentProtection(
   root: Element
 ) : IParsedContentProtection|undefined {
-    let schemeIdUri : string|undefined;
-    let value : string|undefined;
-    let KID : string|undefined;
-    let pssh : string|undefined;
-    for (let i = 0; i < root.attributes.length; i++) {
-      const attribute = root.attributes[i];
+  let schemeIdUri : string|undefined;
+  let value : string|undefined;
+  let keyId : string|undefined;
+  for (let i = 0; i < root.attributes.length; i++) {
+    const attribute = root.attributes[i];
 
-      switch (attribute.name) {
-        case "schemeIdUri":
-          schemeIdUri = attribute.value;
-          break;
-        case "value":
-          value = attribute.value;
-          break;
-        case "cenc:default_KID":
-          KID = attribute.value.toString().split("-").join("").toUpperCase();
-      }
+    switch (attribute.name) {
+      case "schemeIdUri":
+        schemeIdUri = attribute.value;
+        break;
+      case "value":
+        value = attribute.value;
+        break;
+      case "cenc:default_KID":
+        keyId = attribute.value.toString().split("-").join("").toUpperCase();
     }
+  }
 
-    for (let i = 0; i < root.childElementCount; i++) {
-      const child = root.children[i];
-      if (child.nodeName === "cenc:pssh" && child.textContent) {
-        pssh = atob(child.textContent);
-      }
-    }
+  // TODO Take systemId from PSSH?
+  // for (let i = 0; i < root.childElementCount; i++) {
+  //   const child = root.children[i];
+  //   if (child.nodeName === "cenc:pssh" && child.textContent) {
+  //     pssh = atob(child.textContent);
+  //   }
+  // }
 
-    return {
-      schemeIdUri,
-      value,
-      KID,
-      pssh,
-    };
+  return {
+    schemeIdUri,
+    value,
+    keyId,
+  };
 }
